feat: Change the managed to use an external store
