+++ packages/src/context/index.ts
@@ -1,2 +1,1 @@
 export { OverlayProvider } from './provider';
-export { useOverlayList } from './context';
