+++ packages/src/event.ts
@@ -1,25 +1,36 @@
-import { type OverlayContextValue, type OverlayControllerComponent } from './context/context';
+import { type OverlayControllerComponent } from './context/provider';
+import { dispatchOverlay } from './context/store';
 import { randomId } from './utils';
-import { createUseExternalEvents } from './utils/create-use-external-events';
 
-type OverlayEvent = Omit<OverlayContextValue, 'overlayList'>;
-
-type OverlayOpenOptions = {
+type OpenOverlayOptions = {
   overlayId?: string;
 };
 
-export const [useOverlayEvent, createEvent] = createUseExternalEvents<OverlayEvent>('overlay-kit');
-
-const open = (controller: OverlayControllerComponent, options?: OverlayOpenOptions) => {
+function open(controller: OverlayControllerComponent, options?: OpenOverlayOptions) {
   const overlayId = options?.overlayId ?? randomId();
-  const dispatchOpenEvent = createEvent('open');
 
-  dispatchOpenEvent({ controller, overlayId });
+  dispatchOverlay({
+    type: 'ADD',
+    overlay: {
+      id: overlayId,
+      isOpen: false,
+      controller: controller,
+    },
+  });
+
   return overlayId;
-};
-const close = createEvent('close');
-const unmount = createEvent('unmount');
-const closeAll = createEvent('closeAll');
-const unmountAll = createEvent('unmountAll');
+}
+function close(id: string) {
+  dispatchOverlay({ type: 'CLOSE', overlayId: id });
+}
+function unmount(id: string) {
+  dispatchOverlay({ type: 'REMOVE', overlayId: id });
+}
+function closeAll() {
+  dispatchOverlay({ type: 'CLOSE_ALL' });
+}
+function unmountAll() {
+  dispatchOverlay({ type: 'REMOVE_ALL' });
+}
 
 export const overlay = { open, close, unmount, closeAll, unmountAll };
